requirements.txt
absl-py==0.13.0
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
entrypoints==0.3
fastai==2.4
fastcore==1.3.20
fastprogress==1.0.0
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.2.0
imutils==0.5.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
MultiApp==1.0
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.4
pydeck==0.6.2
pydicom==2.1.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.15.0
smart-open==5.1.0
smmap==4.0.0
spacy==3.0.6
spacy-legacy==3.0.6
srsly==2.4.1
streamlit==0.84.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-io==0.19.0
tensorflow-io-gcs-filesystem==0.19.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
thinc==8.0.7
threadpoolctl==2.1.0
tifffile==2021.6.14
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchcam==0.2.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
